SP finalSScore saving

+++ app/src/main/java/com/example/fishmania/PlayerFish.java
@@ -3,7 +3,7 @@ package com.example.fishmania;
 public class PlayerFish extends Fish {
 
     public PlayerFish() {
-        super(5, GameLevel.EASY);
+        super((Integer) 5, GameLevel.EASY);
     }
 
     //setFishValue(getFishValue()+1);
